Program to calculate shooting percentage based on distance
This program will make a lot of API requests to the NHL API to get
the shooting percentage of players at all distances. It is currently
set up to show the 2017-2018 regular season, however, it can be
modified for any year and for preseason/postseason. It is quite
slow because of all the requests, so expect to wait a few hours.